pyproject.toml
[project]
name = "contacts-app"
version = "0.1.0"
description = "App for managing contacts"
authors = [
    {name = "Valera",email = "v.mischenko90@gmail.com"}
]
readme = "README.md"

[tool.poetry]
packages = [{include = "*", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
sqlalchemy = "^2.0.43"
asyncpg = "^0.30.0"
alembic = "^1.16.5"
fastapi = {extras = ["standard"], version = "^0.116.1"}
pydantic-settings = "^2.10.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cloudinary = "^1.44.1"
libgravatar = "^1.0.4"
fastapi-mail = "^1.5.0"
slowapi = "^0.1.9"
redis = "^6.4.0"
pytest = "^8.4.2"
aiosqlite = "^0.21.0"
pytest-cov = "^7.0.0"
pytest-asyncio = "^1.2.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
